// src/enumFaces/interfaces.ts
export interface FilterSettings {
    fullTime: boolean;
    partTime: boolean;
    contract: boolean;
    freelance: boolean;
    lead: boolean;
    expert: boolean;
    senior: boolean;
    midRegular: boolean;
    junior: boolean;
    intern: boolean;
    remote: boolean;
    partRemote: boolean;
    onSite: boolean;
    salary: number;
}