// src/enumFaces/enums.ts
export enum ButtonType {
   Primary = 'primary',
   Text = 'text',
}

export enum FilterTypes {
        fullTime = "fullTime",
        partTime = "partTime",
        contract = "contract",
        freelance = "freelance",
        lead = "lead",
        expert = "expert",
        senior = "senior",
        midRegular = "midRegular",
        junior = "junior",
        intern = "intern",
        remote = "remote",
        partRemote = "partRemote",
        onSite = "onSite",
        salary = "salary",
}